requirements.txt
langchain
langchain-openai
langchain-community
langgraph
psycopg2-binary
chromadb
pypdf
numexpr
sqlparse
python-dotenv